pyproject.toml
[project]
name = "fasthtml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.0.1",
    "python-fasthtml>=0.12.30",
]
